# datadogconnector/api_resources.go
/*
 * Opal Custom App Connector API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datadogconnector

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddResourceUser - 
func AddResourceUser(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetResource - 
func GetResource(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetResourceAccessLevels - 
func GetResourceAccessLevels(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetResourceUsers - 
func GetResourceUsers(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetResources - 
func GetResources(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// RemoveResourceUser - 
func RemoveResourceUser(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}