# datadogconnector/api_users.go
/*
 * Opal Custom App Connector API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datadogconnector

import (
	"net/http"

	"github.com/DataDog/datadog-api-client-go/v2/api/datadog"
	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV2"
	"github.com/gin-gonic/gin"
)

// GetUsers -
func GetUsers(c *gin.Context) {
	ctx := datadog.NewDefaultContext(c.Request.Context())
	configuration := datadog.NewConfiguration()
	apiClient := datadog.NewAPIClient(configuration)
	api := datadogV2.NewUsersApi(apiClient)
	resp, _, err := api.ListUsers(ctx, *datadogV2.NewListUsersOptionalParameters())
	if err != nil {
		c.JSON(http.StatusInternalServerError, &Error{
			Code:    http.StatusInternalServerError,
			Message: err.Error(),
		})
		return
	}

	var users []User
	for _, user := range resp.GetData() {
		var email string
		if user.GetAttributes().Email != nil {
			email = *user.GetAttributes().Email
		}
		users = append(users, User{
			Id:    user.GetId(),
			Email: email,
		})
	}

	c.JSON(http.StatusOK, UsersResponse{
		Users:      users,
		NextCursor: nil,
	})
}