# datadogconnector/model_resource_access_levels_response.go
/*
 * Opal Custom App Connector API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datadogconnector

type ResourceAccessLevelsResponse struct {

	// The cursor with which to continue pagination if additional result pages exist.
	NextCursor *string `json:"next_cursor,omitempty"`

	AccessLevels []AccessLevel `json:"access_levels"`
}